change tensor dtypes for data

+++ data.py
@@ -156,10 +156,12 @@ class CorpusSC(Dataset):
                 return_tensors="pt",
             )
             input_ids = ids["input_ids"]
-            attention_mask = ids["attention_mask"]
-            token_type_ids = ids["token_type_ids"]
+            attention_mask = ids["attention_mask"].to(torch.long)
+            token_type_ids = ids["token_type_ids"].to(torch.long)
 
-            labels = torch.tensor([self.label_dict[label] for label in label_list])
+            labels = torch.tensor(
+                [self.label_dict[label] for label in label_list], dtype=torch.uint8
+            )
         else:
             with open(path, encoding="utf-8") as f:
                 data = [json.loads(jline) for jline in f.readlines()]
